Better parsing for "'%s' imported but unused".
This new regex tries to parse the actual name of the module being
imported, instead of passing the entire message. It only extracts the
module name, discarding the rest of the package hierarchy (which is
useless for finding what to underline in ST anyway).

e.g.
```python
from package.nestedpackage.modulex import y
```

> 'package.nestedpackage.modulex.y' imported but unused

`near='y'`