Allow different python versions and enable virtualenv.
Unless cmd is defined with `@python`, SublimeLinter will always use ST's
python, which is version 3. By using `@python` we can choose the version
on a per-project basis.

Pull request #41 has removed this, which IMO was a mistake.

Also, without `check_version = True`, SublimeLinter will not honor the
python version defined in the settings (e.g. .sublimelinterrc), and will
always default to the system version. This is crucial for usage in
projects with a defined virtualenv, where the "@python" setting is set
to the virutalenv's binary.